refactor websearch into a service
