+++ packages/core/src/generation.ts
@@ -49,9 +49,10 @@ import {
     //VerifiableInferenceProvider,
     TelemetrySettings,
     TokenizerType,
+    IWebSearchService,
+    SearchOptions,
 } from "./types.ts";
 import { fal } from "@fal-ai/client";
-import { tavily } from "@tavily/core";
 
 type Tool = CoreTool<any, any>;
 type StepResult = AIStepResult<any>;
@@ -1725,23 +1726,26 @@ export const generateCaption = async (
 };
 
 export const generateWebSearch = async (
-    query: string,
+    data: {
+        query: string
+        options?: SearchOptions
+    },
     runtime: IAgentRuntime
 ): Promise<SearchResponse> => {
     try {
-        const apiKey = runtime.getSetting("TAVILY_API_KEY") as string;
-        if (!apiKey) {
-            throw new Error("TAVILY_API_KEY is not set");
+        const { query, options } = data;
+        const webSearchService =
+        runtime.getService<IWebSearchService>(
+            ServiceType.WEB_SEARCH
+        );
+
+        if (!webSearchService) {
+            throw new Error("Web search service not found");
         }
-        const tvly = tavily({ apiKey });
-        const response = await tvly.search(query, {
-            includeAnswer: true,
-            maxResults: 3, // 5 (default)
-            topic: "general", // "general"(default) "news"
-            searchDepth: "basic", // "basic"(default) "advanced"
-            includeImages: false, // false (default) true
-        });
+
+        const response = await webSearchService.search(query, runtime, options);
         return response;
+
     } catch (error) {
         elizaLogger.error("Error:", error);
     }
